refactor: streamline paths and utilities
