+++ astro.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig, fontProviders } from 'astro/config'
+import { fileURLToPath } from 'node:url'
 
 import icon from 'astro-icon'
 
@@ -30,7 +31,12 @@ export default defineConfig({
   ],
 
   vite: {
-    plugins: [tailwindcss()],
+    plugins: [...tailwindcss()],
+    resolve: {
+      alias: {
+        '@': fileURLToPath(new URL('./src', import.meta.url)),
+      },
+    },
   },
 
   experimental: {
