+++ src/components/commission/featured.astro
@@ -1,10 +1,10 @@
 ---
 import { Image } from 'astro:assets'
-import data from '../../data/commission/commission.json'
-import priority from '../../data/commission/priorityList.json'
+import data from '@/data/commission/commission.json'
+import priority from '@/data/commission/priorityList.json'
 import IllustratorInfo from './illustrator-info.astro'
-import { prepareData } from '../../utils/commission'
-import { imageImports } from '../../data/imageImports'
+import { prepareData } from '@/lib/commission'
+import { imageImports } from '@/data/imageImports'
 
 interface Item {
   fileName: string
