+++ src/components/commission/list.astro
@@ -1,8 +1,8 @@
 ---
 import { Image } from 'astro:assets'
-import data from '../../data/commission/commission.json'
+import data from '@/data/commission/commission.json'
 import IllustratorInfo from './illustrator-info.astro'
-import { imageImports } from '../../data/imageImports'
+import { imageImports } from '@/data/imageImports'
 
 interface Props {
   character: string
