+++ src/components/header.astro
@@ -1,6 +1,6 @@
 ---
 import { Image } from 'astro:assets'
-import menus from '../collections/menu.json'
+import menus from '@/collections/menu.json'
 import { Icon } from 'astro-icon/components'
 
 import LogoLight from '../../public/assets/logo/logo_light.svg'
