+++ src/components/home/friends.astro
@@ -1,5 +1,5 @@
 ---
-import friends from '../../data/friends.json'
+import friends from '@/data/friends.json'
 ---
 
 <section class="mx-auto mb-14 w-full max-w-4xl px-7 lg:px-0">
