+++ src/components/home/writings.astro
@@ -1,6 +1,6 @@
 ---
-import Button from '../button.astro'
-import PostsLoop from '../posts-loop.astro'
+import Button from '@/components/button.astro'
+import PostsLoop from '@/components/posts-loop.astro'
 
 const feed = 'https://blog.aozaki.cc/rss.xml'
 ---
