+++ src/components/photo-gallery.astro
@@ -1,7 +1,7 @@
 ---
 import { Image } from 'astro:assets'
-import photo from '../collections/photo.json'
-import { imageImports } from '../data/imageImports'
+import photo from '@/collections/photo.json'
+import { imageImports } from '@/data/imageImports'
 
 interface Props {
   title: string
