+++ src/data/friends.json
@@ -11,7 +11,7 @@
   },
   {
     "name": "Outvi V",
-    "link":"https://outv.im",
+    "link": "https://outv.im",
     "favicon": "outv.png"
   },
   {
