+++ src/pages/about.astro
@@ -1,9 +1,9 @@
 ---
 import { Image } from 'astro:assets'
-// import experiences from '../collections/experiences.json'
-// import AboutExperience from '../components/about-experience.astro'
-import PageHeading from '../components/page-heading.astro'
-import Layout from '../layouts/main.astro'
+// import experiences from '@/collections/experiences.json'
+// import AboutExperience from '@/components/about-experience.astro'
+import PageHeading from '@/components/page-heading.astro'
+import Layout from '@/layouts/main.astro'
 
 import Banner from '../../public/assets/images/banner.jpg'
 import MyCar from '../../public/assets/images/car.jpg'
