+++ src/pages/commission.astro
@@ -1,9 +1,9 @@
 ---
-import Layout from '../layouts/commission.astro'
-import Featured from '../components/commission/featured.astro'
-import List from '../components/commission/list.astro'
-import commissions from '../data/commission/commission.json'
-import charPriority from '../data/commission/characterPriority.json'
+import Layout from '@/layouts/commission.astro'
+import Featured from '@/components/commission/featured.astro'
+import List from '@/components/commission/list.astro'
+import commissions from '@/data/commission/commission.json'
+import charPriority from '@/data/commission/characterPriority.json'
 
 interface Item {
   Character: string
