+++ src/pages/index.astro
@@ -1,11 +1,11 @@
 ---
-// import Button from '../components/button.astro'
-// import Photography from '../components/home/photography.astro'
+// import Button from '@/components/button.astro'
+// import Photography from '@/components/home/photography.astro'
 import { Image } from 'astro:assets'
-import Separator from '../components/home/separator.astro'
-import Writings from '../components/home/writings.astro'
-import Friends from '../components/home/friends.astro'
-import Layout from '../layouts/main.astro'
+import Separator from '@/components/home/separator.astro'
+import Writings from '@/components/home/writings.astro'
+import Friends from '@/components/home/friends.astro'
+import Layout from '@/layouts/main.astro'
 ---
 
 <Layout title="Aozaki's blog">
