+++ src/pages/post/[slug].astro
@@ -1,6 +1,6 @@
 ---
 import { getCollection } from 'astro:content'
-import PostLayout from '../../layouts/post.astro' // ← 确认路径
+import PostLayout from '@/layouts/post.astro' // ← 确认路径
 
 // ➊ 生成静态路径
 export async function getStaticPaths() {
