+++ src/pages/posts/[page].astro
@@ -1,7 +1,7 @@
 ---
-import PageHeading from '../../components/page-heading.astro'
-import PostsLoop from '../../components/posts-loop.astro'
-import Layout from '../../layouts/main.astro'
+import PageHeading from '@/components/page-heading.astro'
+import PostsLoop from '@/components/posts-loop.astro'
+import Layout from '@/layouts/main.astro'
 import { getCollection } from 'astro:content'
 
 import { PER_PAGE } from './index.astro'
